Implemented .gitignore, eliminated redundant main entries, corrected Makefile configurations, and standardized source file extensions
