--- src/vm/main.cpp
@@ -1,3 +1,0 @@
-int main() {
-    
-}
