+++ src/vm/vm.hh
